requirements.txt
python-telegram-bot
requests
python-dotenv
apscheduler
SQLAlchemy
psycopg2-binary
